#70 ✅ test: 도메인 실패 이벤트 Kakfa 전송 및 aop 단위 테스트 코드 추가
