+++ snail-member/src/main/kotlin/com/bockerl/snailmember/common/event/BaseFileEvent.kt
@@ -1,9 +1,9 @@
 package com.bockerl.snailmember.common.event
 
 import com.bockerl.snailmember.file.command.domain.aggregate.enums.FileTargetType
-import com.bockerl.snailmember.file.command.domain.aggregate.event.FileCreatedEvent
 import com.bockerl.snailmember.file.command.domain.aggregate.event.FileDeletedEvent
-import com.bockerl.snailmember.file.command.domain.aggregate.event.GatheringFileCreatedEvent
+import com.bockerl.snailmember.file.command.domain.aggregate.event.FileEvent
+import com.bockerl.snailmember.file.command.domain.aggregate.event.GatheringFileEvent
 import com.fasterxml.jackson.annotation.JsonSubTypes
 import com.fasterxml.jackson.annotation.JsonTypeInfo
 
@@ -16,16 +16,16 @@ import com.fasterxml.jackson.annotation.JsonTypeInfo
     property = "eventType",
 )
 @JsonSubTypes(
-    JsonSubTypes.Type(value = FileCreatedEvent::class, name = "FILE"),
-    JsonSubTypes.Type(value = FileCreatedEvent::class, name = "FILE_CREATED"),
-    JsonSubTypes.Type(value = GatheringFileCreatedEvent::class, name = "GATHERING_FILE"),
+    JsonSubTypes.Type(value = FileEvent::class, name = "FILE"),
+    JsonSubTypes.Type(value = FileEvent::class, name = "FILE_CREATED"),
+    JsonSubTypes.Type(value = GatheringFileEvent::class, name = "GATHERING_FILE"),
     JsonSubTypes.Type(value = FileDeletedEvent::class, name = "FILE_DELETED"),
 )
-interface BaseFileCreatedEvent {
-    val fileName: String
-    val fileUrl: String
-    val fileType: String
+interface BaseFileEvent {
+//    val fileName: String
+//    val fileUrl: String
+//    val fileType: String
     val fileTargetType: FileTargetType
     val fileTargetId: String
-    val memberId: String
+//    val memberId: String
 }
