+++ snail-member/src/main/kotlin/com/bockerl/snailmember/common/event/DomainName.kt
@@ -1,0 +1,15 @@
+package com.bockerl.snailmember.common.event
+
+enum class DomainName {
+    ACTIVITY_AREA,
+    BOARD,
+    BOARD_COMMENT,
+    BOARD_COMMENT_LIKE,
+    BOARD_LIKE,
+    BOARD_RECOMMENT,
+    BOARD_RECOMMENT_LIKE,
+    FILE,
+    OUTBOX,
+    SEARCH,
+    MEMBER,
+}
