+++ snail-member/src/main/kotlin/com/bockerl/snailmember/file/command/domain/aggregate/event/FileDeletedEvent.kt
@@ -1,9 +1,9 @@
 package com.bockerl.snailmember.file.command.domain.aggregate.event
 
-import com.bockerl.snailmember.common.event.BaseFileCreatedEvent
+import com.bockerl.snailmember.common.event.BaseFileEvent
 import com.bockerl.snailmember.file.command.domain.aggregate.enums.FileTargetType
 
 data class FileDeletedEvent(
-    val fileTargetType: FileTargetType,
-    val fileTargetId: String,
-) : BaseFileCreatedEvent
+    override val fileTargetType: FileTargetType,
+    override val fileTargetId: String,
+) : BaseFileEvent
