+++ snail-member/src/main/kotlin/com/bockerl/snailmember/file/command/domain/aggregate/event/GatheringFileEvent.kt
@@ -1,14 +1,14 @@
 package com.bockerl.snailmember.file.command.domain.aggregate.event
 
-import com.bockerl.snailmember.common.event.BaseFileCreatedEvent
+import com.bockerl.snailmember.common.event.BaseFileEvent
 import com.bockerl.snailmember.file.command.domain.aggregate.enums.FileTargetType
 
-data class FileCreatedEvent(
+class GatheringFileEvent(
     val fileName: String,
     val fileUrl: String,
     val fileType: String,
-    val fileTargetType: FileTargetType,
-    val fileTargetId: String,
+    override val fileTargetType: FileTargetType,
+    override val fileTargetId: String,
     val memberId: String,
 //    override val fileName: String,
 //    override val fileUrl: String,
@@ -16,4 +16,5 @@ data class FileCreatedEvent(
 //    override val fileTargetType: FileTargetType,
 //    override val fileTargetId: String,
 //    override val memberId: String,
-) : BaseFileCreatedEvent
+    val gatheringId: String,
+) : BaseFileEvent
