+++ snail-member/src/main/kotlin/com/bockerl/snailmember/infrastructure/aop/Logging.kt
@@ -1,0 +1,5 @@
+package com.bockerl.snailmember.infrastructure.aop
+
+@Target(AnnotationTarget.FUNCTION)
+@Retention(AnnotationRetention.RUNTIME)
+annotation class Logging
