+++ snail-member/src/main/kotlin/com/bockerl/snailmember/member/command/domain/service/CommandMemberServiceImpl.kt
@@ -1,11 +1,11 @@
 package com.bockerl.snailmember.member.command.domain.service
 
-import com.bockerl.snailmember.area.command.application.service.CommandAreaService
 import com.bockerl.snailmember.common.exception.CommonException
 import com.bockerl.snailmember.common.exception.ErrorCode
 import com.bockerl.snailmember.file.command.application.dto.CommandFileDTO
 import com.bockerl.snailmember.file.command.application.service.CommandFileService
 import com.bockerl.snailmember.file.command.domain.aggregate.enums.FileTargetType
+import com.bockerl.snailmember.infrastructure.aop.Logging
 import com.bockerl.snailmember.infrastructure.outbox.dto.OutboxDTO
 import com.bockerl.snailmember.infrastructure.outbox.enums.EventType
 import com.bockerl.snailmember.infrastructure.outbox.service.OutboxService
@@ -27,7 +27,6 @@ import java.time.LocalDateTime
 @Service
 class CommandMemberServiceImpl(
     private val memberRepository: MemberRepository,
-    private val activityAreaService: CommandAreaService,
     private val commandFileService: CommandFileService,
     private val outboxService: OutboxService,
     private val eventPublisher: ApplicationEventPublisher,
@@ -36,6 +35,7 @@ class CommandMemberServiceImpl(
     private val logger = KotlinLogging.logger {}
 
     @Transactional
+    @Logging
     override fun putLastAccessTime(
         email: String,
         ipAddress: String,
@@ -55,6 +55,7 @@ class CommandMemberServiceImpl(
                         memberId = member.formattedId,
                         userAgent = userAgent,
                         ipAddress = ipAddress,
+                        idemPotencyKey = idempotencyKey,
                     )
                 eventPublisher.publishEvent(event)
                 // outBox를 통한 이벤트 처리
@@ -70,12 +71,12 @@ class CommandMemberServiceImpl(
             }.onSuccess {
                 logger.info { "마지막 로그인 시간 업데이트 성공 - email: $email" }
             }.onFailure {
-                logger.info { "마지막 로그인 시간 업데이트 실패, memberId: ${member.formattedId}" }
                 throw CommonException(ErrorCode.INTERNAL_SERVER_ERROR, "로그인 시간 업데이트 실패 - email: $email")
             }.getOrThrow()
     }
 
     @Transactional
+    @Logging
     override fun patchProfile(
         memberId: String,
         requestDTO: ProfileRequestDTO,
@@ -106,14 +107,6 @@ class CommandMemberServiceImpl(
             )
         // logging을 위한 비동기 리스너 이벤트 처리
         eventPublisher.publishEvent(event)
-        memberRepository
-            .runCatching {
-                save(member)
-            }.onSuccess {
-                logger.info { "프로필 사진 제외 회원 정보 수정 성공" }
-            }.onFailure {
-                throw CommonException(ErrorCode.INTERNAL_SERVER_ERROR, "프로플 사진 제외 회원 정보 수정 실패")
-            }
         file?.let {
             logger.info { "프로필 사진 수정 시작" }
             val commandFileDTO =
@@ -124,24 +117,23 @@ class CommandMemberServiceImpl(
                     idempotencyKey = idempotencyKey,
                 )
             // 기존 프사가 없다면 생성, 있다면 삭제 후 수정 호출
-            if (member.memberPhoto.isBlank()) {
-                val fileUrl = commandFileService.createSingleFile(file, commandFileDTO)
-                logger.info { "프로필 사진 저장 성공 - fileUrl: $fileUrl" }
-                member.apply { memberPhoto = fileUrl }
-            } else {
-                val fileUrl = commandFileService.updateProfileImage(file, commandFileDTO)
-                logger.info { "프로필 사진 저장 성공 - fileUrl: $fileUrl" }
-                member.apply { memberPhoto = fileUrl }
-            }
-            memberRepository
-                .runCatching {
-                    save(member)
-                }.onSuccess {
-                    logger.info { "프로필 사진 수정 성공" }
-                }.onFailure {
-                    throw CommonException(ErrorCode.INTERNAL_SERVER_ERROR, "프로필 사진 수정 실패")
-                }.getOrThrow()
+            val fileUrl =
+                if (member.memberPhoto.isBlank()) {
+                    commandFileService.createSingleFile(file, commandFileDTO)
+                } else {
+                    commandFileService.updateProfileImage(file, commandFileDTO)
+                }
+            logger.info { "프로필 사진 저장 성공 - fileUrl: $fileUrl" }
+            member.memberPhoto = fileUrl
         }
+        memberRepository
+            .runCatching {
+                save(member)
+            }.onSuccess {
+                logger.info { "회원 정보 및 프로필 사진 수정 성공" }
+            }.onFailure {
+                throw CommonException(ErrorCode.INTERNAL_SERVER_ERROR, "회원 정보 수정 실패")
+            }.getOrThrow()
         val jsonPayLoad = objectMapper.writeValueAsString(event)
         val outBox =
             OutboxDTO(
@@ -154,6 +146,7 @@ class CommandMemberServiceImpl(
     }
 
     @Transactional
+    @Logging
     override fun deleteMember(
         memberId: String,
         idempotencyKey: String,
@@ -178,7 +171,6 @@ class CommandMemberServiceImpl(
                 logger.info { "멤버 탈퇴 성공, memberId: $memberId" }
             }.onFailure {
                 logger.info { "멤버 탈퇴 실패, memberId: $memberId" }
-                // 로그 전송 및 보상 트랜잭션 예상
                 throw CommonException(ErrorCode.INTERNAL_SERVER_ERROR)
             }.getOrThrow()
         // 회원 삭제 이벤트 아웃박스 생성
