+++ snail-member/src/main/kotlin/com/bockerl/snailmember/member/command/domain/service/GoogleOauthServiceImpl.kt
@@ -2,13 +2,14 @@ package com.bockerl.snailmember.member.command.domain.service
 
 import com.bockerl.snailmember.common.exception.CommonException
 import com.bockerl.snailmember.common.exception.ErrorCode
+import com.bockerl.snailmember.infrastructure.aop.Logging
 import com.bockerl.snailmember.infrastructure.config.TransactionalConfig
 import com.bockerl.snailmember.infrastructure.outbox.dto.OutboxDTO
 import com.bockerl.snailmember.infrastructure.outbox.enums.EventType
 import com.bockerl.snailmember.infrastructure.outbox.service.OutboxService
-import com.bockerl.snailmember.member.client.LineAuthClient
-import com.bockerl.snailmember.member.command.application.dto.response.LinePayloadDTO
-import com.bockerl.snailmember.member.command.application.service.LineOauth2Service
+import com.bockerl.snailmember.member.client.GoogleAuthClient
+import com.bockerl.snailmember.member.command.application.dto.response.GooglePayloadDTO
+import com.bockerl.snailmember.member.command.application.service.GoogleOauth2Service
 import com.bockerl.snailmember.member.command.config.Oauth2LoginProperties
 import com.bockerl.snailmember.member.command.domain.aggregate.entity.Member
 import com.bockerl.snailmember.member.command.domain.aggregate.entity.enums.Gender
@@ -25,58 +26,56 @@ import org.springframework.context.ApplicationEventPublisher
 import org.springframework.security.core.authority.SimpleGrantedAuthority
 import org.springframework.security.core.userdetails.UserDetails
 import org.springframework.stereotype.Service
+import org.springframework.transaction.annotation.Transactional
 import java.time.LocalDate
 import java.util.Base64
 import java.util.UUID
 
 @Service
-class LineOauth2ServiceImpl(
-    private val lineAuthClient: LineAuthClient,
+class GoogleOauthServiceImpl(
     private val memberRepository: MemberRepository,
     private val loginProperties: Oauth2LoginProperties,
+    private val googleAuthClient: GoogleAuthClient,
     private val jwtUtils: Oauth2JwtUtils,
     private val objectMapper: ObjectMapper,
     private val eventPublisher: ApplicationEventPublisher,
     private val outboxService: OutboxService,
-) : LineOauth2Service {
+) : GoogleOauth2Service {
     private val logger = KotlinLogging.logger {}
 
-    override fun lineLogin(code: String) =
+    override fun googleLogin(code: String) =
         TransactionalConfig.run {
             // 코드 기반으로 요청을 보낸 뒤 id token만 추출
-            val idToken = requestTokenFromLine(code)
+            val idToken = requestTokenFromGoogle(code)
             // 유저 정보 디코딩
             val customMember = decodeUserInfoFromToken(idToken) as CustomMember
             jwtUtils.generateJwtResponse(customMember)
         }
 
-    private fun requestTokenFromLine(code: String): String {
-        logger.info { "라인 코드 기반 인증 토큰 요청 시작" }
-        logger.info { "clientId: ${loginProperties.lineClientId}" }
-        logger.info { "redirectUri: ${loginProperties.lineRedirectUri}" }
+    private fun requestTokenFromGoogle(code: String): String {
+        logger.info { "구글 코드 기반 인증 토큰 요청 시작" }
+        logger.info { "clientId: ${loginProperties.googleClientId}" }
+        logger.info { "redirectUri: ${loginProperties.googleRedirectUri}" }
         logger.info { "code: $code" }
-        logger.info { "clientSecret: ${loginProperties.lineClientSecret}" }
+        logger.info { "clientSecret: ${loginProperties.googleClientSecret}" }
         return try {
-            val formData =
-                buildString {
-                    append("grant_type=authorization_code")
-                    append("&code=$code")
-                    append("&redirect_uri=${loginProperties.lineRedirectUri}")
-                    append("&client_id=${loginProperties.lineClientId}")
-                    append("&client_secret=${loginProperties.lineClientSecret}")
-                }
-            logger.info { "라인 토큰 요청을 위한 formData: $formData" }
-            val response = lineAuthClient.getAccessToken(formData)
-            logger.info { "라인으로부터 돌아온 토큰 responseDTO: $response" }
+            val response =
+                googleAuthClient.getAccessToken(
+                    clientId = loginProperties.googleClientId,
+                    redirectUri = loginProperties.googleRedirectUri,
+                    code = code,
+                    clientSecret = loginProperties.googleClientSecret,
+                )
+            logger.info { "구글로부터 돌아온 토큰 responseDTO: $response" }
             response.idToken
         } catch (e: Exception) {
-            logger.error(e) { "라인 Access Token 요청 실패" }
+            logger.error(e) { "구글 Access Token 요청 실패" }
             throw e
         }
     }
 
     private fun decodeUserInfoFromToken(idToken: String): UserDetails {
-        logger.info { "라인 ID 토큰으로 유저 정보 디코딩 시작" }
+        logger.info { "구글 ID 토큰으로 유저 정보 디코딩 시작" }
 
         return try {
             // JWT 토큰의 payload 부분 추출 및 디코딩
@@ -90,50 +89,55 @@ class LineOauth2ServiceImpl(
             // JSON 파싱
             val jsonObject = ObjectMapper().readTree(decodedString)
 
-            // id 토큰 제공자가 라인인지 확인
-            if (jsonObject["iss"].asText() != "https://access.line.me"
+            // id 토큰 제공자가 구글인지 확인
+            if (jsonObject["iss"].asText() != "https://accounts.google.com" &&
+                jsonObject["iss"].asText() != "accounts.google.com"
             ) {
-                logger.error { "line id token 제공자가 line이 아닌 에러" }
-                throw CommonException(ErrorCode.LINE_AUTH_ERROR)
+                logger.error { "구글 id token 제공자가 google이 아닌 에러" }
+                throw CommonException(ErrorCode.GOOGLE_AUTH_ERROR)
             }
 
-            // sub 필드가 라인 회원 번호
-            val lineId = jsonObject["sub"].asText()
-            logger.info { "라인 제공 고유 id: $lineId" }
+            // sub 필드가 구글 회원 번호
+            val googleId = jsonObject["sub"].asText()
+            logger.info { "구글 제공 고유 id: $googleId" }
 
             // 이메일 생성
-            val email = "$lineId@line.com"
+            val email = "$googleId@google.com"
 
-            // 라인 payload 데이터 구성
-            val lineResponse =
-                LinePayloadDTO(
-                    id = lineId,
+            // 구글 payload 데이터 구성
+            val googleResponse =
+                GooglePayloadDTO(
+                    id = googleId,
                     name = jsonObject["name"]?.asText(),
                 )
 
-            logger.info { "디코딩된 라인 계정 유저 정보: $lineResponse" }
+            logger.info { "디코딩된 구글 계정 유저 정보: $googleResponse" }
 
             // 기존 회원 조회 또는 새 회원 생성
             val member =
                 memberRepository.findMemberByMemberEmailAndMemberStatusNot(email, MemberStatus.ROLE_DELETED)
-                    ?: createNewLineMember(email, lineResponse)
+                    ?: createNewGoogleMember(email, googleResponse)
             if (member.memberStatus == MemberStatus.ROLE_BLACKLIST) {
-                logger.warn { "라인 블랙 리스트 멤버가 로그인 - email: $email" }
+                logger.warn { "구글 블랙 리스트 멤버가 로그인 - email: $email" }
                 throw CommonException(ErrorCode.BLACK_LIST_ROLE)
             }
             val authority = listOf(SimpleGrantedAuthority(member.memberStatus.toString()))
             CustomMember(member, authority)
-        } catch (e: Exception) {
-            logger.error(e) { "라인 ID 토큰 디코딩 실패" }
+        } catch (e: IllegalArgumentException) {
+            logger.error(e) { "구글 ID 토큰 디코딩 실패" }
             throw IllegalArgumentException("Invalid ID token", e)
+        } catch (e: CommonException) {
+            throw e
         }
     }
 
-    private fun createNewLineMember(
+    @Transactional
+    @Logging
+    fun createNewGoogleMember(
         email: String,
-        lineResponse: LinePayloadDTO,
+        googleResponse: GooglePayloadDTO,
     ): Member {
-        val newLineMember =
+        val newGoogleMember =
             Member(
                 memberEmail = email,
                 memberPhoneNumber = UUID.randomUUID().toString(),
@@ -142,38 +146,37 @@ class LineOauth2ServiceImpl(
                 memberRegion = "",
                 memberLanguage = Language.KOR,
                 memberGender = Gender.UNKNOWN,
-                memberNickname = lineResponse.name ?: UUID.randomUUID().toString(),
+                memberNickname = googleResponse.name ?: UUID.randomUUID().toString(),
                 memberBirth = LocalDate.now(),
                 memberPassword = UUID.randomUUID().toString(),
-                signupPath = SignUpPath.LINE,
+                signupPath = SignUpPath.Google,
                 selfIntroduction = "",
             )
 
-        logger.info { "새로 생성되는 라인 계정 멤버: $newLineMember" }
+        logger.info { "새로 생성되는 구글 계정 멤버: $newGoogleMember" }
         memberRepository
             .runCatching {
-                save(newLineMember)
+                save(newGoogleMember)
             }.onSuccess {
-                logger.info { "라인 계정 새 멤버 저장 성공" }
+                logger.info { "구글 계정 새 멤버 저장 성공" }
             }.onFailure {
-                logger.error { "라인 계정 새 멤버 저장 실패, member: $newLineMember, message: ${it.message}" }
+                logger.error { "구글 계정 새 멤버 저장 실패, member: $newGoogleMember, message: ${it.message}" }
                 throw CommonException(ErrorCode.INTERNAL_SERVER_ERROR)
             }.getOrThrow()
-
         // outbox 이벤트 발행(회원 생성)
         val event =
             MemberCreateEvent(
-                memberId = newLineMember.formattedId,
-                memberEmail = newLineMember.memberEmail,
-                memberPhoneNumber = newLineMember.memberPhoneNumber,
-                memberStatus = newLineMember.memberStatus,
-                memberRegion = newLineMember.memberRegion,
-                memberGender = newLineMember.memberGender,
-                memberNickname = newLineMember.memberNickname,
-                memberPhoto = newLineMember.memberPhoto,
-                memberBirth = newLineMember.memberBirth,
-                memberLanguage = newLineMember.memberLanguage,
-                signUpPath = SignUpPath.LINE,
+                memberId = newGoogleMember.formattedId,
+                memberEmail = newGoogleMember.memberEmail,
+                memberPhoneNumber = newGoogleMember.memberPhoneNumber,
+                memberStatus = newGoogleMember.memberStatus,
+                memberRegion = newGoogleMember.memberRegion,
+                memberGender = newGoogleMember.memberGender,
+                memberNickname = newGoogleMember.memberNickname,
+                memberPhoto = newGoogleMember.memberPhoto,
+                memberBirth = newGoogleMember.memberBirth,
+                memberLanguage = newGoogleMember.memberLanguage,
+                signUpPath = SignUpPath.Google,
             )
 
         // logging을 위한 비동기 리스너 이벤트 처리
@@ -182,13 +185,13 @@ class LineOauth2ServiceImpl(
         val jsonPayLoad = objectMapper.writeValueAsString(event)
         val outBox =
             OutboxDTO(
-                aggregateId = newLineMember.formattedId,
+                aggregateId = newGoogleMember.formattedId,
                 eventType = EventType.MEMBER,
                 payload = jsonPayLoad,
                 // 회원가입은 1번만 발생하므로, pk를 멱등키로
-                idempotencyKey = newLineMember.formattedId,
+                idempotencyKey = newGoogleMember.formattedId,
             )
         outboxService.createOutbox(outBox)
-        return newLineMember
+        return newGoogleMember
     }
 }
